% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{florida_vaccine_coverage}
\alias{florida_vaccine_coverage}
\title{Kindergarten Vaccination Coverage for Florida Counties}
\format{
A data frame with 748 rows and 3 variables:
\describe{
  \item{County}{The name of the county in Florida, plus a statewide "Florida" total.}
  \item{Year}{The calendar year of the report, from 2016 to 2024.}
  \item{Coverage_Rate}{The proportion (0 to 1) of kindergarten students with complete immunizations.}
}
}
\source{
Florida Department of Health, Bureau of Immunization.
}
\usage{
florida_vaccine_coverage
}
\description{
A dataset containing the percentage of kindergarten students who have completed
required immunizations for school entry in Florida counties.
}
\keyword{datasets}
