% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_coverage_history}
\alias{plot_coverage_history}
\title{Plot Historical Vaccination Coverage Rate}
\usage{
plot_coverage_history(county_name, save_data_to = NULL)
}
\arguments{
\item{county_name}{A character vector of county names to plot.}

\item{save_data_to}{Optional file path to save the plot's data.}
}
\value{
A ggplot object.
}
\description{
Plot Historical Vaccination Coverage Rate
}
\examples{
plot_coverage_history(county_name = c("Florida", "Miami-Dade", "Liberty"))
}
