% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_calculators.R
\name{prob_major_outbreak}
\alias{prob_major_outbreak}
\title{Calculate the probability of a major outbreak}
\usage{
prob_major_outbreak(Re)
}
\arguments{
\item{Re}{The effective reproduction number.}
}
\value{
The probability of a major outbreak (1 - 1/Re).
}
\description{
Calculate the probability of a major outbreak
}
\examples{
prob_major_outbreak(Re = 2.05)
}
