% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_calculators.R
\name{prob_sec_case}
\alias{prob_sec_case}
\title{Calculate the probability of at least one secondary infection}
\usage{
prob_sec_case(Re)
}
\arguments{
\item{Re}{The effective reproduction number.}
}
\value{
The probability of at least one secondary infection.
}
\description{
Calculate the probability of at least one secondary infection
}
\examples{
prob_sec_case(Re = 2.05)
}
